test: make the correctness suite honest and green
Each TEST_CASE is now its own ctest (label `correctness`), so one crash no
longer hides the rest; perf benchmarks move to the `perf` label. CI and
nightly run the whole correctness suite.

Root-caused the failures the single-process run was hiding:
- brass_throw is an asm entry point on x64 (MinGW clobbered rdi before the
  callee-saved capture); the MSVC stub reserved too little stack.
- TLABs are detached from their old HostGC on rebind and re-pointed on move,
  fixing the thread-exit use-after-free.
- COFF REL32/PLT32 inline addends are measured from the field end, and
  loaded sections carry their real alignment.
- The bronze corpus AOT bridge defines bronze_arg_at; the end-to-end AOT
  link test runs under the MSVC env and fails instead of skipping.

Public headers are compiled with consumer-style flags so a header warning
(C4324 in jit_exec.hpp broke bronze) fails brass's own build. MSVC and
MinGW builds are warning-clean; computed-goto dispatch no longer leaks the
coroutine DeoptFrame/vector.

+++ tests/differential/test_pic_differential.cpp
@@ -175,7 +175,7 @@ func accessProperty(%0: dynamic) -> f64 {
     double expected_vals[3] = {123.0, 456.0, 789.0};
 
     for (int i = 0; i < 3000; ++i) {
-        int idx = rng() % 3;
+        int idx = static_cast<int>(rng() % 3);
         double result = access_fn(reinterpret_cast<uint64_t>(arr[idx]));
         CHECK_EQ(result, expected_vals[idx]);
     }
